split watcher tests into two files
